requirements.txt
python-dotenv
openai
transformers
SpeechRecognition
requests
pydub
ffmpeg-python
google-cloud-speech